requirements.txt
pandas
pyarrow
python-dateutil
